+++ models/navWeb.go
@@ -3,7 +3,7 @@ package models
 type NavWeb struct {
 	Authentication bool
 	AccountID      int
-	TypeAccess     int
+	TypeAccess     uint8
 	IsPremmium     bool
 	AccountName    string
 	MyPlayers      []Players
